refactor: changing folder style
