+++ lib/collections/decorators/collection.decorator.ts
@@ -1,8 +1,8 @@
 /**
  * Interface defining schema options that can be passed to `@Collection()` decorator.
  */
-import { MetadataStorage } from '../storages/metadata.storage';
-import { CollectionUtils } from './collection.utils';
+import { MetadataStorage } from '../../storages/metadata.storage';
+import { CollectionUtils } from '../utils/collection.utils';
 import { FirestoreDataConverter } from '@google-cloud/firestore';
 import { defaultConverter } from '@google-cloud/firestore/build/src/types';
