+++ lib/collections/errors/collection-conflict.error.ts
