+++ lib/collections/errors/collection-not-defined.error.ts
