+++ lib/repositories/batches/errors/write-batch-not-declared.error.ts
