+++ lib/repositories/batches/utils/write-batch.utils.ts
