+++ lib/repositories/decorators/inject-repository.decorator.ts
@@ -1,5 +1,5 @@
 import { Inject, Type } from '@nestjs/common';
-import { getRepositoryToken } from '../firestore.constants';
+import { getRepositoryToken } from '../../firestore.constants';
 
 export const InjectRepository = (documentType: Type): ParameterDecorator =>
   Inject(getRepositoryToken(documentType));
