+++ lib/repositories/errors/id-required-for-op.error.ts
@@ -1,0 +1,5 @@
+export class IdRequiredForOpError extends Error {
+  constructor() {
+    super('ID is required for this operation.');
+  }
+}
