+++ lib/repositories/queries/dto/firestore-document.dto.ts
