+++ lib/repositories/transactions/errors/transaction-not-declared.error.ts
