+++ lib/repositories/transactions/transaction.provider.ts
@@ -5,7 +5,7 @@ import {
   UpdateData,
   WithFieldValue,
 } from '@google-cloud/firestore';
-import { FirestoreDocument } from '../queries/firestore-document.dto';
+import { FirestoreDocument } from '../queries/dto/firestore-document.dto';
 
 export class Transaction {
   private firestoreTransaction: FirestoreTransaction;
